+++ crates/core/src/message/encoder/test_encoder.rs
@@ -13,11 +13,6 @@ fn test_encode_decode() {
     let result1: Vec<u8> = encoded1.decode().unwrap();
     assert_eq!(test1, result1);
 
-    let test2 = "abc";
-    let encoded2 = test2.encode().unwrap();
-    let result2: String = encoded2.decode().unwrap();
-    assert_eq!(test2, result2);
-
     let test3 = String::from("abc");
     let encoded3 = test3.encode().unwrap();
     let result3: String = encoded3.decode().unwrap();